requirements.txt
pandas
sklearn
numpy
FastAPI
uvicorn[standard]
sentence-transformers
tqdm
h5py
weaviate-client